task 1/requirements.txt
praw
tweepy
gspread
oauth2client
python-dotenv
SpeechRecognition
pyaudio